+++ examples/python_trace_writer/__init__.py
@@ -1,0 +1,3 @@
+from .agentlens_trace import AgentLensRun, init_workspace, trace_llm_call, write_trace
+
+__all__ = ["AgentLensRun", "init_workspace", "trace_llm_call", "write_trace"]